+++ DemoApp/StreamChat/Components/CustomAttachments/LocationAttachment/LocationAttachmentViewDelegate.swift
@@ -4,16 +4,52 @@
 
 import StreamChat
 import StreamChatUI
+import UIKit
 
 protocol LocationAttachmentViewDelegate: ChatMessageContentViewDelegate {
-    func didTapOnLocationAttachment(
-        _ attachment: ChatMessageLocationAttachment
+    func didTapOnLocation(
+        _ location: SharedLocation
+    )
+
+    func didTapOnStopSharingLocation(
+        _ location: SharedLocation
     )
 }
 
 extension DemoChatMessageListVC: LocationAttachmentViewDelegate {
-    func didTapOnLocationAttachment(_ attachment: ChatMessageLocationAttachment) {
-        let mapViewController = LocationDetailViewController(locationAttachment: attachment)
+    func didTapOnLocation(_ location: SharedLocation) {
+        let messageController = client.messageController(
+            cid: location.channelId,
+            messageId: location.messageId
+        )
+        showDetailViewController(messageController: messageController)
+    }
+
+    func didTapOnStopSharingLocation(_ location: SharedLocation) {
+        client
+            .messageController(cid: location.channelId, messageId: location.messageId)
+            .stopLiveLocationSharing { result in
+                switch result {
+                case .success:
+                    break
+                case .failure(let error):
+                    self.presentAlert(
+                        title: "Could not stop sharing location",
+                        message: error.localizedDescription
+                    )
+                }
+            }
+    }
+
+    private func showDetailViewController(messageController: ChatMessageController) {
+        let mapViewController = LocationDetailViewController(
+            messageController: messageController
+        )
+        if UIDevice.current.userInterfaceIdiom == .pad {
+            let nav = UINavigationController(rootViewController: mapViewController)
+            navigationController?.present(nav, animated: true)
+            return
+        }
         navigationController?.pushViewController(mapViewController, animated: true)
     }
 }