+++ DemoApp/StreamChat/Components/DemoChatMessageActionsVC.swift
@@ -33,6 +33,11 @@ final class DemoChatMessageActionsVC: ChatMessageActionsVC {
             actions.append(messageDebugActionItem())
         }
 
+        let hasLocationAttachments = message?.sharedLocation != nil
+        if hasLocationAttachments {
+            actions.removeAll(where: { $0 is EditActionItem })
+        }
+
         return actions
     }