+++ DemoApp/StreamChat/StreamChatWrapper+DemoApp.swift
@@ -9,10 +9,6 @@ import StreamChatUI
 extension StreamChatWrapper {
     // Instantiates chat client
     func setUpChat() {
-        if AppConfig.shared.demoAppConfig.isLocationAttachmentsEnabled {
-            Components.default.mixedAttachmentInjector.register(.location, with: LocationAttachmentViewInjector.self)
-        }
-
         // Set the log level
         LogConfig.level = StreamRuntimeCheck.logLevel ?? .warning
         LogConfig.formatters = [
@@ -26,7 +22,6 @@ extension StreamChatWrapper {
         if client == nil {
             client = ChatClient(config: config)
         }
-        client?.registerAttachment(LocationAttachmentPayload.self)
 
         // L10N
         let localizationProvider = Appearance.default.localizationProvider