+++ Sources/StreamChat/Database/DTOs/MessageDTO.swift
@@ -88,6 +88,9 @@ class MessageDTO: NSManagedObject {
     @NSManaged var draftReply: MessageDTO?
     @NSManaged var isDraft: Bool
 
+    @NSManaged var location: SharedLocationDTO?
+    @NSManaged var isActiveLiveLocation: Bool
+    
     @NSManaged var reminder: MessageReminderDTO?
 
     /// If the message is sent by the current user, this field
@@ -123,6 +126,13 @@ class MessageDTO: NSManagedObject {
 
         if let channel = channel, self.cid != channel.cid {
             self.cid = channel.cid
+        }
+
+        if let locationEndAt = location?.endAt?.bridgeDate {
+            let isActiveLiveLocation = locationEndAt > Date()
+            if isActiveLiveLocation != self.isActiveLiveLocation {
+                self.isActiveLiveLocation = isActiveLiveLocation
+            }
         }
 
         // Manually mark the channel as dirty to trigger the entity update and give the UI a chance
@@ -617,7 +627,59 @@ class MessageDTO: NSManagedObject {
         ])
         return try load(request, context: context)
     }
-    
+
+    /// Fetches all active location messages in a channel or all channels of the current user.
+    /// If `channelId` is nil, it will fetch all messages independent of the channel.
+    static func currentUserActiveLiveLocationMessagesFetchRequest(
+        currentUserId: UserId,
+        channelId: ChannelId?
+    ) -> NSFetchRequest<MessageDTO> {
+        let request = NSFetchRequest<MessageDTO>(entityName: MessageDTO.entityName)
+        MessageDTO.applyPrefetchingState(to: request)
+        // Hard coded limit for now. 10 live locations messages at the same should be more than enough.
+        request.fetchLimit = 10
+        request.sortDescriptors = [NSSortDescriptor(
+            keyPath: \MessageDTO.createdAt,
+            ascending: true
+        )]
+        var predicates: [NSPredicate] = [
+            .init(format: "isActiveLiveLocation == YES"),
+            .init(format: "user.id == %@", currentUserId),
+            .init(format: "localMessageStateRaw == nil")
+        ]
+        if let channelId {
+            predicates.append(.init(format: "channel.cid == %@", channelId.rawValue))
+        }
+        request.predicate = NSCompoundPredicate(andPredicateWithSubpredicates: predicates)
+        return request
+    }
+
+    /// Fetches all active location messages in any given channel and from every user.
+    static func activeLiveLocationMessagesFetchRequest() -> NSFetchRequest<MessageDTO> {
+        let request = NSFetchRequest<MessageDTO>(entityName: MessageDTO.entityName)
+        MessageDTO.applyPrefetchingState(to: request)
+        request.fetchLimit = 25
+        request.sortDescriptors = [NSSortDescriptor(
+            keyPath: \MessageDTO.createdAt,
+            ascending: true
+        )]
+        var predicates: [NSPredicate] = [
+            .init(format: "isActiveLiveLocation == YES"),
+            .init(format: "localMessageStateRaw == nil")
+        ]
+        request.predicate = NSCompoundPredicate(andPredicateWithSubpredicates: predicates)
+        return request
+    }
+
+    static func loadCurrentUserActiveLiveLocationMessages(
+        currentUserId: UserId,
+        channelId: ChannelId?,
+        context: NSManagedObjectContext
+    ) throws -> [MessageDTO] {
+        let request = currentUserActiveLiveLocationMessagesFetchRequest(currentUserId: currentUserId, channelId: channelId)
+        return try load(request, context: context)
+    }
+
     static func loadReplies(
         from fromIncludingDate: Date,
         to toIncludingDate: Date,
@@ -680,6 +742,7 @@ extension NSManagedObjectContext: MessageDatabaseSession {
         skipPush: Bool,
         skipEnrichUrl: Bool,
         poll: PollPayload?,
+        location: NewLocationInfo?,
         restrictedVisibility: [UserId],
         extraData: [String: RawJSON]
     ) throws -> MessageDTO {
@@ -691,7 +754,8 @@ extension NSManagedObjectContext: MessageDatabaseSession {
             throw ClientError.ChannelDoesNotExist(cid: cid)
         }
 
-        let message = MessageDTO.loadOrCreate(id: messageId ?? .newUniqueId, context: self, cache: nil)
+        let id = messageId ?? .newUniqueId
+        let message = MessageDTO.loadOrCreate(id: id, context: self, cache: nil)
 
         // We make `createdDate` 0.1 second bigger than Channel's most recent message
         // so if the local time is not in sync, the message will still appear in the correct position
@@ -733,6 +797,23 @@ extension NSManagedObjectContext: MessageDatabaseSession {
 
         if let poll {
             message.poll = try? savePoll(payload: poll, cache: nil)
+        }
+
+        if let location, let currentUser, let deviceId = currentUser.currentDevice?.id {
+            message.location = try? saveLocation(
+                payload: .init(
+                    channelId: cid.rawValue,
+                    messageId: id,
+                    userId: currentUser.user.id,
+                    latitude: location.latitude,
+                    longitude: location.longitude,
+                    createdAt: Date(),
+                    updatedAt: Date(),
+                    endAt: location.endAt,
+                    createdByDeviceId: deviceId
+                ),
+                cache: nil
+            )
         }
 
         message.attachments = Set(
@@ -949,6 +1030,10 @@ extension NSManagedObjectContext: MessageDatabaseSession {
             }
         }
 
+        if let location = payload.location {
+            dto.location = try saveLocation(payload: location, cache: cache)
+        }
+
         let user = try saveUser(payload: payload.user)
         dto.user = user
 
@@ -1005,7 +1090,7 @@ extension NSManagedObjectContext: MessageDatabaseSession {
             }
         )
         dto.attachments = attachments
-        
+
         if let poll = payload.poll {
             let pollDto = try savePoll(payload: poll, cache: cache)
             dto.poll = pollDto
@@ -1591,6 +1676,14 @@ extension MessageDTO {
             pinExpires: pinExpires?.bridgeDate,
             pollId: poll?.id,
             restrictedVisibility: restrictedVisibilityArray,
+            location: location.map {
+                .init(
+                    latitude: $0.latitude,
+                    longitude: $0.longitude,
+                    endAt: $0.endAt?.bridgeDate,
+                    createdByDeviceId: $0.deviceId
+                )
+            },
             extraData: extraData
         )
     }
@@ -1706,6 +1799,7 @@ private extension ChatMessage {
         }
         
         let poll = try? dto.poll?.asModel()
+        let location = try? dto.location?.asModel()
 
         let currentUserReactions: Set<ChatMessageReaction>
         let isSentByCurrentUser: Bool
@@ -1803,7 +1897,8 @@ private extension ChatMessage {
                 remindAt: $0.remindAt?.bridgeDate,
                 createdAt: $0.createdAt.bridgeDate,
                 updatedAt: $0.updatedAt.bridgeDate
-            ) }
+            ) },
+            sharedLocation: location
         )
 
         if let transformer = chatClientConfig?.modelsTransformer {