+++ TestTools/StreamChatTestTools/Mocks/Models + Extensions/ChatMessage_Mock.swift
@@ -51,7 +51,8 @@ public extension ChatMessage {
         textUpdatedAt: Date? = nil,
         poll: Poll? = nil,
         draftReply: DraftMessage? = nil,
-        reminder: MessageReminderInfo? = nil
+        reminder: MessageReminderInfo? = nil,
+        sharedLocation: SharedLocation? = nil
     ) -> Self {
         .init(
             id: id,
@@ -93,7 +94,8 @@ public extension ChatMessage {
             poll: poll,
             textUpdatedAt: textUpdatedAt,
             draftReply: draftReply,
-            reminder: reminder
+            reminder: reminder,
+            sharedLocation: sharedLocation
         )
     }
 }