+++ TestTools/StreamChatTestTools/TestData/DummyData/MessageAttachmentPayload.swift
@@ -1,5 +1,5 @@
 //
-// Copyright © 2025 Stream.io Inc. All rights reserved.
+// Copyright 2025 Stream.io Inc. All rights reserved.
 //
 
 import Foundation