+++ Tests/StreamChatTests/Database/DTOs/ChannelDTO_Tests.swift
@@ -397,6 +397,7 @@ final class ChannelDTO_Tests: XCTestCase {
             Assert.willBeEqual(payload.channel.config.createdAt, loadedChannel.config.createdAt)
             Assert.willBeEqual(payload.channel.config.updatedAt, loadedChannel.config.updatedAt)
             Assert.willBeEqual(payload.channel.config.messageRemindersEnabled, loadedChannel.config.messageRemindersEnabled)
+            Assert.willBeEqual(payload.channel.config.sharedLocationsEnabled, loadedChannel.config.sharedLocationsEnabled)
 
             // Own Capabilities
             Assert.willBeEqual(payload.channel.ownCapabilities, ["join-channel", "delete-channel"])
@@ -1112,7 +1113,8 @@ final class ChannelDTO_Tests: XCTestCase {
             pinnedMessages: [],
             channelReads: [currentUserChannelReadPayload],
             isHidden: false,
-            draft: nil
+            draft: nil,
+            activeLiveLocations: []
         )
 
         let unreadMessages = 5
@@ -1169,7 +1171,8 @@ final class ChannelDTO_Tests: XCTestCase {
             pinnedMessages: [],
             channelReads: [],
             isHidden: nil,
-            draft: nil
+            draft: nil,
+            activeLiveLocations: []
         )
         try database.writeSynchronously { session in
             try session.saveChannel(payload: channelPayload)
@@ -1537,7 +1540,8 @@ final class ChannelDTO_Tests: XCTestCase {
                 quotedMessage: nil,
                 parentId: nil,
                 parentMessage: nil
-            )
+            ),
+            activeLiveLocations: [.dummy(latitude: 10, longitude: 10)]
         )
 
         // WHEN
@@ -1552,6 +1556,8 @@ final class ChannelDTO_Tests: XCTestCase {
         XCTAssertEqual(draftMessage.id, draftMessagePayload.id)
         XCTAssertEqual(draftMessage.text, draftMessagePayload.text)
         XCTAssertEqual(draftMessage.extraData, draftMessagePayload.extraData)
+        XCTAssertEqual(channel.activeLiveLocations.first?.latitude, 10)
+        XCTAssertEqual(channel.activeLiveLocations.first?.longitude, 10)
     }
 
     func test_saveChannel_whenDraftMessageIsNil_removesExistingDraft() throws {
@@ -1588,7 +1594,8 @@ final class ChannelDTO_Tests: XCTestCase {
                 quotedMessage: nil,
                 parentId: nil,
                 parentMessage: nil
-            )
+            ),
+            activeLiveLocations: []
         )
 
         try database.writeSynchronously { session in
@@ -1613,7 +1620,8 @@ final class ChannelDTO_Tests: XCTestCase {
             pinnedMessages: [],
             channelReads: [],
             isHidden: nil,
-            draft: nil
+            draft: nil,
+            activeLiveLocations: []
         )
 
         try database.writeSynchronously { session in
@@ -1664,7 +1672,8 @@ final class ChannelDTO_Tests: XCTestCase {
                 quotedMessage: quotedMessagePayload,
                 parentId: nil,
                 parentMessage: nil
-            )
+            ),
+            activeLiveLocations: []
         )
 
         // WHEN
@@ -1720,7 +1729,8 @@ final class ChannelDTO_Tests: XCTestCase {
                 quotedMessage: nil,
                 parentId: parentMessagePayload.id,
                 parentMessage: parentMessagePayload
-            )
+            ),
+            activeLiveLocations: []
         )
 
         // WHEN