+++ CHANGELOG.md
@@ -3,7 +3,24 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
-### 🔄 Changed
+## StreamChat
+### ✅ Added
+- Add `ChatMessageController.partialUpdateMessage()` [#3531](https://github.com/GetStream/stream-chat-swift/pull/3531)
+- Add Location Sharing Support [#3531](https://github.com/GetStream/stream-chat-swift/pull/3531)
+  - Add `ChatMessage.sharedLocation`
+  - Add `ChatMessageController.stopLiveLocationSharing()`
+  - Add `ChatChannelController`:
+    - `sendStaticLocation()`
+    - `startLiveLocationSharing()`
+  - Add `CurrentChatUserController`:
+    - `loadActiveLiveLocationMessages()`
+    - `updateLiveLocation()`
+  - Add `CurrentChatUserControllerDelegate`:
+    - `didStartSharingLiveLocation()`
+    - `didStopSharingLiveLocation()`
+    - `didChangeActiveLiveLocationMessages()`
+    - `didFailToUpdateLiveLocation()`
+
 
 # [4.80.0](https://github.com/GetStream/stream-chat-swift/releases/tag/4.80.0)
 _June 17, 2025_